requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
distro==1.9.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
openai==2.9.0
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sniffio==1.3.1
sqlparse==0.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==6.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
